requirements.txt

pandas>=2.0.0
numpy>=1.24.0
spacy>=3.7.0
torch>=2.2.0
tqdm>=4.66.0
scikit-learn>=1.3.0
pyyaml>=6.0.1